refactor entity files into entities package
