+++ src/main/java/DisplayView/Table.java
@@ -1,6 +1,5 @@
 package DisplayView;
 
-import DisplayZoom.DataPoint;
 import DisplayZoom.DataPointMap;
 
 import java.util.ArrayList;
