+++ src/main/java/DisplayView/VisualUnit.java
@@ -1,5 +1,5 @@
 package DisplayView;
-//import DisplayZoom.DataPoint;
+//import entities.DataPoint;
 //import DisplayZoom.DataPointMap;
 import java.util.Date;
