+++ src/main/java/DisplayZoom/DataPointManager.java
@@ -1,5 +1,7 @@
 package DisplayZoom;
 
+import entities.DataPoint;
+
 /**
  * DataPointManager class
  */
