+++ src/main/java/exercise/ExerciseManager.java
@@ -2,6 +2,8 @@ package exercise;
 
 // Use Case
 
+import entities.Exercise;
+
 public class ExerciseManager {
 
     /**
