+++ src/main/java/user_creator/UserAccountList.java
@@ -1,5 +1,7 @@
 package user_creator;
 
+import entities.User;
+
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
